[PROJ-91] refactor: PropertyManage 위치 변경
